Make mvar compile against CE3
Remove MVar and rename MVar 2 as we are not maintaining binary compatibility
Remove Concurrent mvar as we don't have the Concurrent/Async distinction
any more where one supports cancellation and the other doesn't
