+++ concurrent/shared/src/main/scala/cats/effect/concurrent/MVar.scala
@@ -17,11 +17,10 @@
 package cats.effect
 package concurrent
 
-import cats.effect.kernel.{Sync, Concurrent, Async}
-import cats.effect.concurrent.MVar.{TransformedMVar, TransformedMVar2}
-import cats.effect.internals.{MVarAsync, MVarConcurrent}
+import cats.effect.kernel.{Async, Sync}
+import cats.effect.concurrent.MVar.TransformedMVar
+import cats.effect.internals.MVarAsync
 import cats.~>
-import com.github.ghik.silencer.silent
 
 /**
  * @define mVarDescription A mutable location, that is either empty or contains a value of type `A`.
@@ -38,15 +37,15 @@ import com.github.ghik.silencer.silent
  *  - [[read]] which reads the current value without touching it,
  *    assuming there is one, or otherwise it waits until a value
  *    is made available via `put`
- *  - `tryRead` returns a variable if it exists. Implemented in the successor [[MVar2]]
- *  - `swap` takes a value, replaces it and returns the taken value. Implemented in the successor [[MVar2]]
+ *  - `tryRead` returns a variable if it exists. Implemented in the successor [[MVar]]
+ *  - `swap` takes a value, replaces it and returns the taken value. Implemented in the successor [[MVar]]
  *  - [[isEmpty]] returns true if currently empty
  *
  * The `MVar` is appropriate for building synchronization
  * primitives and performing simple inter-thread communications.
  * If it helps, it's similar with a `BlockingQueue(capacity = 1)`,
  * except that it doesn't block any threads, all waiting being
- * done asynchronously (via [[Async]] or [[Concurrent]] data types,
+ * done asynchronously (via [[Async]] data types,
  * such as [[IO]]).
  *
  * Given its asynchronous, non-blocking nature, it can be used on
@@ -60,7 +59,6 @@ sealed private[concurrent] trait MVarDocumentation extends Any {}
 /**
  * $mVarDescription
  */
-@deprecated("`MVar` is now deprecated in favour of a new generation `MVar2` with `tryRead` and `swap` support", "2.2.0")
 abstract class MVar[F[_], A] extends MVarDocumentation {
 
   /**
@@ -123,22 +121,6 @@ abstract class MVar[F[_], A] extends MVarDocumentation {
   def read: F[A]
 
   /**
-   * Modify the context `F` using transformation `f`.
-   */
-  def mapK[G[_]](f: F ~> G): MVar[G, A] =
-    new TransformedMVar(this, f)
-}
-
-/**
- * $mVarDescription
- *
- * The `MVar2` is the successor of `MVar` with [[tryRead]] and [[swap]]. It was implemented separately only to maintain
- * binary compatibility with `MVar`.
- */
-@silent("deprecated")
-abstract class MVar2[F[_], A] extends MVar[F, A] {
-
-  /**
    * Replaces a value in MVar and returns the old value.
 
    * @param newValue is a new value
@@ -158,17 +140,17 @@ abstract class MVar2[F[_], A] extends MVar[F, A] {
   /**
    * Modify the context `F` using transformation `f`.
    */
-  override def mapK[G[_]](f: F ~> G): MVar2[G, A] =
-    new TransformedMVar2(this, f)
+  def mapK[G[_]](f: F ~> G): MVar[G, A] =
+    new TransformedMVar(this, f)
 }
 
 /** Builders for [[MVar]]. */
 object MVar {
 
   /**
-   * Builds an [[MVar]] value for `F` data types that are [[Concurrent]].
+   * Builds an [[MVar]] value for `F` data types that are [[Async]].
    *
-   * Due to `Concurrent`'s capabilities, the yielded values by [[MVar.take]]
+   * Due to `Async`'s capabilities, the yielded values by [[MVar.take]]
    * and [[MVar.put]] are cancelable.
    *
    * This builder uses the
@@ -187,7 +169,7 @@ object MVar {
    *
    * @see [[of]]and [[empty]]
    */
-  def apply[F[_]](implicit F: Concurrent[F]): ApplyBuilders[F] =
+  def apply[F[_]](implicit F: Async[F]): ApplyBuilders[F] =
     new ApplyBuilders[F](F)
 
   /**
@@ -198,23 +180,7 @@ object MVar {
    * @param F is a [[Concurrent]] constraint, needed in order to
    *        describe cancelable operations
    */
-  def empty[F[_], A](implicit F: Concurrent[F]): F[MVar2[F, A]] =
-    F.delay(MVarConcurrent.empty)
-
-  /**
-   * Creates a non-cancelable `MVar` that starts as empty.
-   *
-   * The resulting `MVar` has non-cancelable operations.
-   *
-   * WARN: some `Async` data types, like [[IO]], can be cancelable,
-   * making `uncancelable` values unsafe. Such values are only useful
-   * for optimization purposes, in cases where the use case does not
-   * require cancellation or in cases in which an `F[_]` data type
-   * that does not support cancellation is used.
-   *
-   * @see [[empty]] for creating cancelable MVars
-   */
-  def uncancelableEmpty[F[_], A](implicit F: Async[F]): F[MVar2[F, A]] =
+  def empty[F[_], A](implicit F: Async[F]): F[MVar[F, A]] =
     F.delay(MVarAsync.empty)
 
   /**
@@ -229,61 +195,32 @@ object MVar {
    * @param F is a [[Concurrent]] constraint, needed in order to
    *        describe cancelable operations
    */
-  def of[F[_], A](initial: A)(implicit F: Concurrent[F]): F[MVar2[F, A]] =
-    F.delay(MVarConcurrent(initial))
-
-  /**
-   * Creates a non-cancelable `MVar` that's initialized to an `initial`
-   * value.
-   *
-   * The resulting `MVar` has non-cancelable operations.
-   *
-   * WARN: some `Async` data types, like [[IO]], can be cancelable,
-   * making `uncancelable` values unsafe. Such values are only useful
-   * for optimization purposes, in cases where the use case does not
-   * require cancellation or in cases in which an `F[_]` data type
-   * that does not support cancellation is used.
-   *
-   * @see [[of]] for creating cancelable MVars
-   */
-  def uncancelableOf[F[_], A](initial: A)(implicit F: Async[F]): F[MVar2[F, A]] =
+  def of[F[_], A](initial: A)(implicit F: Async[F]): F[MVar[F, A]] =
     F.delay(MVarAsync(initial))
 
   /**
    * Like [[of]] but initializes state using another effect constructor
    */
-  def in[F[_], G[_], A](initial: A)(implicit F: Sync[F], G: Concurrent[G]): F[MVar2[G, A]] =
-    F.delay(MVarConcurrent(initial))
+  def in[F[_], G[_], A](initial: A)(implicit F: Sync[F], G: Async[G]): F[MVar[G, A]] =
+    F.delay(MVarAsync(initial))
 
   /**
    * Like [[empty]] but initializes state using another effect constructor
    */
-  def emptyIn[F[_], G[_], A](implicit F: Sync[F], G: Concurrent[G]): F[MVar2[G, A]] =
-    F.delay(MVarConcurrent.empty)
-
-  /**
-   * Like [[uncancelableOf]] but initializes state using another effect constructor
-   */
-  def uncancelableIn[F[_], G[_], A](initial: A)(implicit F: Sync[F], G: Async[G]): F[MVar2[G, A]] =
-    F.delay(MVarAsync(initial))
-
-  /**
-   * Like [[uncancelableEmpty]] but initializes state using another effect constructor
-   */
-  def uncancelableEmptyIn[F[_], G[_], A](implicit F: Sync[F], G: Async[G]): F[MVar2[G, A]] =
+  def emptyIn[F[_], G[_], A](implicit F: Sync[F], G: Async[G]): F[MVar[G, A]] =
     F.delay(MVarAsync.empty)
 
   /**
    * Returned by the [[apply]] builder.
    */
-  final class ApplyBuilders[F[_]](val F: Concurrent[F]) extends AnyVal {
+  final class ApplyBuilders[F[_]](val F: Async[F]) extends AnyVal {
 
     /**
      * Builds an `MVar` with an initial value.
      *
      * @see documentation for [[MVar.of]]
      */
-    def of[A](a: A): F[MVar2[F, A]] =
+    def of[A](a: A): F[MVar[F, A]] =
       MVar.of(a)(F)
 
     /**
@@ -291,22 +228,12 @@ object MVar {
      *
      * @see documentation for [[MVar.empty]]
      */
-    def empty[A]: F[MVar2[F, A]] =
+    def empty[A]: F[MVar[F, A]] =
       MVar.empty(F)
   }
 
   final private[concurrent] class TransformedMVar[F[_], G[_], A](underlying: MVar[F, A], trans: F ~> G)
       extends MVar[G, A] {
-    override def isEmpty: G[Boolean] = trans(underlying.isEmpty)
-    override def put(a: A): G[Unit] = trans(underlying.put(a))
-    override def tryPut(a: A): G[Boolean] = trans(underlying.tryPut(a))
-    override def take: G[A] = trans(underlying.take)
-    override def tryTake: G[Option[A]] = trans(underlying.tryTake)
-    override def read: G[A] = trans(underlying.read)
-  }
-
-  final private[concurrent] class TransformedMVar2[F[_], G[_], A](underlying: MVar2[F, A], trans: F ~> G)
-      extends MVar2[G, A] {
     override def isEmpty: G[Boolean] = trans(underlying.isEmpty)
     override def put(a: A): G[Unit] = trans(underlying.put(a))
     override def tryPut(a: A): G[Boolean] = trans(underlying.tryPut(a))
